Removed non-irq-friendly trace call in irq context
During debugging, the tr_debug call in M2MConnectionHandlerPimpl::socket_event
took an unacceptable amount of time to complete, causing bytes traveling
through the serial connections to be dropped.

This was unnoticable in lwip's network-socket implementation due to not
occuring in an interrupt context, but prevented the esp8266 from working
correctly.